+++ crates/nostr-relay-pool/src/relay/mod.rs
@@ -372,7 +372,7 @@ impl Relay {
 
     /// Get events of filters with custom callback
     #[inline]
-    pub(crate) async fn get_events_of_with_callback<F>(
+    pub(crate) async fn fetch_events_with_callback<F>(
         &self,
         filters: Vec<Filter>,
         timeout: Duration,
@@ -383,29 +383,50 @@ impl Relay {
         F: Future<Output = ()>,
     {
         self.inner
-            .get_events_of_with_callback(filters, timeout, opts, callback)
+            .fetch_events_with_callback(filters, timeout, opts, callback)
             .await
     }
 
-    /// Get events of filters
+    /// Fetch events
     #[inline]
+    pub async fn fetch_events(
+        &self,
+        filters: Vec<Filter>,
+        timeout: Duration,
+        opts: FilterOptions,
+    ) -> Result<Vec<Event>, Error> {
+        self.inner.fetch_events(filters, timeout, opts).await
+    }
+
+    /// Get events of filters
+    #[deprecated(since = "0.36.0", note = "Use `fetch_events` instead")]
     pub async fn get_events_of(
         &self,
         filters: Vec<Filter>,
         timeout: Duration,
         opts: FilterOptions,
     ) -> Result<Vec<Event>, Error> {
-        self.inner.get_events_of(filters, timeout, opts).await
+        self.fetch_events(filters, timeout, opts).await
+    }
+
+    /// Count events
+    #[inline]
+    pub async fn count_events(
+        &self,
+        filters: Vec<Filter>,
+        timeout: Duration,
+    ) -> Result<usize, Error> {
+        self.inner.count_events(filters, timeout).await
     }
 
     /// Count events of filters
-    #[inline]
+    #[deprecated(since = "0.36.0", note = "Use `count_events` instead")]
     pub async fn count_events_of(
         &self,
         filters: Vec<Filter>,
         timeout: Duration,
     ) -> Result<usize, Error> {
-        self.inner.count_events_of(filters, timeout).await
+        self.count_events(filters, timeout).await
     }
 
     /// Negentropy reconciliation