+++ crates/nostr-sdk/examples/gossip.rs
@@ -51,10 +51,7 @@ async fn main() -> Result<()> {
     // Get events
     let filter = Filter::new().author(pubkey).kind(Kind::TextNote).limit(3);
     let events = client
-        .get_events_of(
-            vec![filter],
-            EventSource::both(Some(Duration::from_secs(10))),
-        )
+        .fetch_events(vec![filter], Some(Duration::from_secs(10)))
         .await?;
 
     for event in events.into_iter() {